+++ R/mod_flags.R
@@ -44,9 +44,9 @@ server_flags <- function(id, wells) {
         dplyr::select("well_tag_number", "lithology_from_m", "lithology_to_m",
                       dplyr::starts_with("flag")) %>%
         dplyr::filter(dplyr::if_any(dplyr::starts_with("flag"))) %>%
-        aq_dt() %>%
+        aq_dt(filename = "flags") %>%
         DT::formatStyle(cols, backgroundColor = DT::styleEqual(TRUE, "#f8d7da"))
-    })
+    }, server = FALSE)
 
     output$flags_glossary <- renderTable(bcaquiferdata::flags)