+++ R/utils.R
@@ -66,19 +66,25 @@ fix_range <- function(x) {
 }
 
 
-
-aq_dt <- function(data, minimal = FALSE) {
+# File names - https://stackoverflow.com/a/56276939
+aq_dt <- function(data, filename = NULL, minimal = FALSE) {
   if(minimal) {
     opts <- list(dom = "t")
     ext <- list()
   } else {
+    filename <- paste0(filename, "-", Sys.Date())
     opts <- list(dom = 'Bfrtip',
-                 buttons = c('csv', 'excel', I('colvis')))
+                 buttons = list(
+                   I('colvis'),
+                   list(extend = 'csv', title = filename),
+                   list(extend = 'excel', title = filename)
+                 ))
     ext <- "Buttons"
   }
 
   data %>%
     DT::datatable(
+      rownames = FALSE,
       fillContainer = TRUE,
       options = append(
         list(pageLength = 14, scrollX = TRUE),