+++ NEWS.md
@@ -1,5 +1,5 @@
 # Version dev
-
+* Fix download buttons (#1)
 
 # Version 0.0.3
 * Fix ArcHydro exports